+++ src/acl/DataTypes/aclSubvector.h
@@ -24,9 +24,12 @@
 #ifndef ACLSUBVECTOR_H
 #define ACLSUBVECTOR_H
 
-#include <CL/cl.hpp>
+//#include <CL/cl.hpp>
+// Supply "cl.hpp" with ASL, since it is not present in OpenCL 2.0
+// Remove the file after switching to OpenCL 2.1
+#include "acl/cl.hpp"
 #include "aclArray.h"
-#include "../../aslUtilities.h"
+#include "aslUtilities.h"
 
 ///\todo{Add padding at the end?}