+++ src/acl/Kernels/aclKernel.h
@@ -26,7 +26,10 @@
 
 #include "aclExpressionContainer.h"
 #include "aclKernelConfiguration.h"
-#include <CL/cl.hpp>
+//#include <CL/cl.hpp>
+// Supply "cl.hpp" with ASL, since it is not present in OpenCL 2.0
+// Remove the file after switching to OpenCL 2.1
+#include "acl/cl.hpp"
 
 namespace acl
 {