+++ src/acl/aclUtilities.cxx
@@ -23,7 +23,7 @@
 
 #include "aclUtilities.h"
 #include "Kernels/aclKernelConfigurationTemplates.h"
-#include <acl/aclElementBase.h>
+#include "acl/aclElementBase.h"
 
 #include "aclTypesList.h"
 #include "aclHardware.h"