+++ src/aslUtilities.cxx
@@ -22,7 +22,7 @@
 
 #include <stdexcept>
 #include "aslUtilities.h"
-#include <acl/aclUtilities.h>
+#include "acl/aclUtilities.h"
 
 using namespace std;