+++ NEWS.md
@@ -2,5 +2,7 @@
 
 ## 0.1.5
 
+- supplying cl.hpp, since not provided by OpenCL 2.0
 - bug fixes
 - improved tests
+- improved examples